added `utils/`; moved `utils.py, math.py` to `utils/`
