+++ langchain/document_transformers/embeddings_redundant_filter.py
@@ -5,8 +5,8 @@ import numpy as np
 from pydantic import BaseModel, Field
 
 from langchain.embeddings.base import Embeddings
-from langchain.math_utils import cosine_similarity
 from langchain.schema import BaseDocumentTransformer, Document
+from langchain.utils.math import cosine_similarity
 
 
 class _DocumentWithState(Document):
