+++ langchain/vectorstores/utils.py
@@ -5,7 +5,7 @@ from typing import List
 
 import numpy as np
 
-from langchain.math_utils import cosine_similarity
+from langchain.utils.math import cosine_similarity
 
 
 class DistanceStrategy(str, Enum):
