+++ tests/unit_tests/test_math_utils.py
@@ -4,7 +4,7 @@ from typing import List
 import numpy as np
 import pytest
 
-from langchain.math_utils import cosine_similarity, cosine_similarity_top_k
+from langchain.utils.math import cosine_similarity, cosine_similarity_top_k
 
 
 @pytest.fixture
